Add a --launch option to appcontroller
This will trigger a "fire and forget" mode where we won't bother with
the server sockets. This allows launching multiple apps with appcontroller
without killing the previous one.

This is the preferred approach to launch Wayland applications. We still would
like to use appcontroller (to set up the environment) but killing the previous
app is not desirable. Tooling does not need to be considered now.

Change-Id: I9119d622b12aaf79cfb2e68c95bacd079dffeb52
Reviewed-by: Rainer Keller <rainer.keller@theqtcompany.com>