Add a way to clear specific environment variables
Currently the configured environment variables are only set if they
don't override an existing one. This is done to allow setting an
environment variable in QtCreator.

This is a problem when a specific environment variable is defined
in the root shell e.g. XDG_RUNTIME_DIR, but the actual process should
use another value, specified using the 'env' key.

To solve this a new 'clearEnv' setting gets introduced which makes it
possible to clear specific environment variables before applying the
new environment variables.

Change-Id: I319db570a4e410b8f2716adcec9be882d89f4b1e
Reviewed-by: Robert Griebl <robert.griebl@qt.io>